pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.55",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "python-dotenv>=1.1.1",
]
